Branch 'rename' changing name of directory holding console scripts to avoid reserved word conflict with Py2.
