+++ rcsb/db/scripts/__init__.py
@@ -3,4 +3,4 @@ __docformat__ = "restructuredtext en"
 __author__ = "John Westbrook"
 __email__ = "john.westbrook@rcsb.org"
 __license__ = "Apache 2.0"
-__version__ = "0.16"
+__version__ = "0.17"
