+++ tox.ini
@@ -70,7 +70,7 @@ deps =
     pyflakes
     pylint
 commands =
-    pylint --rcfile=.pylintrc --ignore=rcsb/db/schema/ rcsb/db/ setup.py
+    pylint --rcfile=.pylintrc --ignore=rcsb/mock-data/ rcsb/db/ setup.py
 #
 
 [testenv:coverage-macos]
@@ -92,14 +92,14 @@ commands =
     # pip install mmcif  --no-cache-dir
     # {envpython} setup.py  sdist
     coverage erase
-    coverage run --parallel-mode --omit="rcsb/db/tests-*","rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/db/schema/*" --source rcsb/db -m unittest discover  --start-directory rcsb/db/tests/ --pattern "*Tests.py"
+    coverage run --parallel-mode --omit="rcsb/db/tests-*","rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/mock-data/*" --source rcsb/db -m unittest discover  --start-directory rcsb/db/tests/ --pattern "*Tests.py"
     #
-    coverage run  --parallel-mode --omit="rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/db/schema/*" --source rcsb/db/ -m unittest discover  --start-directory rcsb/db/tests-mongo/ --pattern "*Tests.py"
+    coverage run  --parallel-mode --omit="rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/mock-data/*" --source rcsb/db/ -m unittest discover  --start-directory rcsb/db/tests-mongo/ --pattern "*Tests.py"
     #
-    coverage run  --parallel-mode --omit="rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/db/schema/*" --source rcsb/db/ -m unittest discover  --start-directory rcsb/db/tests-mysql/ --pattern "*Tests.py"
+    coverage run  --parallel-mode --omit="rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/mock-data/*" --source rcsb/db/ -m unittest discover  --start-directory rcsb/db/tests-mysql/ --pattern "*Tests.py"
     coverage combine
     #
-    coverage report --omit="rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/db/schema/*","rcsb/db/tests*","rcsb/db/exec/*",   --show-missing --fail-under=50
+    coverage report --omit="rcsb/db/crate/*","rcsb/db/cockroach/*","rcsb/mock-data/*","rcsb/db/tests*","rcsb/db/scripts/*",   --show-missing --fail-under=50
     - coverage html
     echo "Done with coverage (macos)"
